requirements.txt
streamlit
opencv-python-headless
pandas
numpy
scikit-learn
pyttsx3
Pillow